Adding overridable unload() to Loader class
This patch adds the possibility to manually provide an implementation
for unloading all tables of the benchmark. While the current
implementation uses SQL only, with this patch its possible to override
the method from Loader::unload() and implement own behavior.

The default implementation uses the same behavior as for the loading
of tables via the `makeLoaderImpl` method of the `BenchmarkModule`